Reorganise deck layer files
