+++ src/lib/nested-config/config-tree.ts
@@ -1,0 +1,1 @@
+export type ConfigTree<T> = (ConfigTree<T> | false | null | undefined | T)[];
