+++ src/state/layers/view-layers-flat.ts
@@ -1,6 +1,6 @@
 import { selector } from 'recoil';
 
-import { flattenConfig } from 'lib/config-tree';
+import { flattenConfig } from 'lib/nested-config/flatten-config';
 import { ViewLayer } from 'lib/data-map/view-layers';
 
 import { viewLayersState } from './view-layers';
